feat: add profileview
