+++ 14th-team5-iOS/App/Sources/Presentation/Notification/Reactor/NotifiactionReactor.swift
@@ -1,0 +1,12 @@
+//
+//  NotifiactionReactor.swift
+//  App
+//
+//  Created by 마경미 on 04.01.25.
+//
+
+import ReactorKit
+
+final class NotifiactionReactor: Reactor {
+    
+}
