+++ 14th-team5-iOS/App/Sources/Presentation/Notification/ViewControllers/NotificationViewController.swift
@@ -1,0 +1,12 @@
+//
+//  NotificationViewController.swift
+//  App
+//
+//  Created by 마경미 on 04.01.25.
+//
+
+import Core
+
+final class NotificationViewController: BBNavigationViewController<NotifiactionReactor> {
+    
+}
