+++ 14th-team5-iOS/Core/Sources/Bibbi/BBCommons/BBProfile/BBProfileTextView.swift
@@ -1,0 +1,123 @@
+//
+//  BBProfile.swift
+//  Core
+//
+//  Created by 마경미 on 23.12.24.
+//
+
+import UIKit
+
+import SnapKit
+import RxSwift
+
+class BBProfile: UIView {
+    enum Style {
+        case imageTopTextBottom  // 이미지가 위, 텍스트가 아래 (상하)
+        case imageLeftTextRight  // 이미지가 좌, 텍스트가 우 (좌우)
+    }
+    
+    public struct Configure {
+        let isBirthday: Bool
+        let image: UIImage?
+        let name: String
+        let comment: String?
+        
+        public init(
+            isBirthday: Bool = false,
+            image: UIImage?,
+            name: String,
+            comment: String?
+        ) {
+            self.isBirthday = isBirthday
+            self.image = image
+            self.name = name
+            self.comment = comment
+        }
+    }
+    
+    private let style: Style
+    
+    public var rx: Reactive<BBProfile> {
+           return Reactive(self)
+    }
+    
+    internal let imageView: BBProfileImageView
+    private let textStackView: UIStackView = UIStackView()
+    internal let nameLabel: BBLabel = BBLabel(.body1Regular, textColor: .gray200)
+    internal let commentLabel: BBLabel = BBLabel(.caption, textColor: .gray500)
+    
+    public init(
+        size: BBProfileImageView.Size,
+        style: Style
+    ) {
+        self.style = style
+        self.imageView = .init(size: size)
+        
+        super.init(frame: .zero)
+        
+        setupUI()
+        setupAutoLayout()
+        setupAttributes()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
+extension BBProfile {
+    private func setupUI() {
+        addSubviews(textStackView)
+        textStackView.addArrangedSubviews(nameLabel, commentLabel)
+    }
+    
+    private func setupAutoLayout() {
+        switch style {
+        case .imageTopTextBottom:
+            setupImageTopTextBottomAutoLayout()
+        case .imageLeftTextRight:
+            setupImageLeftTextRightAutoLayout()
+        }
+    }
+    
+    private func setupAttributes() {
+        textStackView.do {
+            $0.axis = .horizontal
+            $0.spacing = 0
+            $0.distribution = .fillProportionally
+        }
+    }
+}
+
+extension BBProfile {
+    private func setupImageTopTextBottomAutoLayout() {
+        textStackView.snp.makeConstraints {
+            $0.leading.equalTo(imageView.snp.trailing).offset(16)
+            $0.directionalVerticalEdges.trailing.equalToSuperview()
+        }
+    }
+    
+    private func setupImageLeftTextRightAutoLayout() {
+        textStackView.snp.makeConstraints {
+            $0.top.equalTo(imageView.snp.bottom).offset(8)
+            $0.directionalHorizontalEdges.bottom.equalToSuperview()
+        }
+    }
+}
+
+extension Reactive where Base: BBProfile {
+    var configure: Binder<BBProfile.Configure> {
+        return Binder(base) { view, config in
+            view.imageView.rx.configure
+                .onNext(.init(
+                    isBirthday: config.isBirthday,
+                    image: config.image,
+                    name: config.name
+                )
+            )
+            view.nameLabel.text = config.name
+            view.commentLabel.text = config.comment
+        }
+    }
+}
+
